+++ backend/src/main/java/com/art/drinkgpt/repositories/IngredientRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IngredientRepository extends JpaRepository<Ingredient, Long> {
+public interface IngredientRepository extends JpaRepository<Ingredient, Integer> {
     Page<Ingredient> findByAvailable(boolean available, Pageable pageable);
 }
